Correct typos and clarify documentation
Documentation in README.md and docs has been updated to correct typographical errors and enhance explanation. The description of the Python script's functionality checking YAML and JSON files is now clearer and more specific in README.md. Additionally, in docs, the use of the "environments" key has been further elaborated for clarity and precision about its application in the web interface.